+++ docs/source/index.rst
@@ -40,7 +40,6 @@ The :mod:`torchaudio` package consists of I/O, popular datasets and common audio
    kaldi_io
    utils
    rnnt_loss
-   tacotron2
 
 
 .. toctree::
