--- docs/source/tacotron2.rst
@@ -1,37 +1,0 @@
-.. role:: hidden
-    :class: hidden-section
-
-torchaudio.prototype.tacotron2
-==============================
-
-.. currentmodule:: torchaudio.prototype.tacotron2
-
-.. note::
-
-    The Tacotron2 model is a prototype feature, see `here <https://pytorch.org/audio>`_ to learn more about the nomenclature.
-    It is only available within the nightlies, and also needs to be imported
-    explicitly using: :code:`from torchaudio.prototype.Tacotron2 import Tacotron2, tacotron2`.
-
-
-Tacotron2
-~~~~~~~~~
-
-.. autoclass:: Tacotron2
-
-  .. automethod:: forward
-
-  .. automethod:: infer
-
-Factory Functions
------------------
-
-tacotron2
----------
-
-.. autofunction:: tacotron2
-
-
-References
-~~~~~~~~~~
-
-.. footbibliography::
