+++ examples/pipeline_tacotron2/inference.py
@@ -11,8 +11,8 @@ import sys
 import torch
 import torchaudio
 import numpy as np
-from torchaudio.prototype.tacotron2 import Tacotron2
-from torchaudio.prototype.tacotron2 import tacotron2 as pretrained_tacotron2
+from torchaudio.models import Tacotron2
+from torchaudio.models import tacotron2 as pretrained_tacotron2
 
 from utils import prepare_input_sequence
 from datasets import InverseSpectralNormalization
@@ -28,7 +28,7 @@ def parse_args():
     r"""
     Parse commandline arguments.
     """
-    from torchaudio.prototype.tacotron2 import _MODEL_CONFIG_AND_URLS as tacotron2_config_and_urls
+    from torchaudio.models.tacotron2 import _MODEL_CONFIG_AND_URLS as tacotron2_config_and_urls
     from torchaudio.models.wavernn import _MODEL_CONFIG_AND_URLS as wavernn_config_and_urls
 
     parser = argparse.ArgumentParser(description=__doc__)
