+++ torchaudio/models/__init__.py
@@ -2,6 +2,7 @@ from .wav2letter import Wav2Letter
 from .wavernn import WaveRNN, wavernn
 from .conv_tasnet import ConvTasNet
 from .deepspeech import DeepSpeech
+from .tacotron2 import Tacotron2, tacotron2
 from .wav2vec2 import (
     Wav2Vec2Model,
     wav2vec2_base,
@@ -20,4 +21,6 @@ __all__ = [
     'wav2vec2_base',
     'wav2vec2_large',
     'wav2vec2_large_lv60k',
+    'Tacotron2',
+    'tacotron2',
 ]
